+++ app/src/main/java/com/boardgamegeek/events/CollectionItemChangedEvent.kt
@@ -1,3 +1,8 @@
 package com.boardgamegeek.events
 
-class CollectionItemChangedEvent(val collectionName: String, val imageUrl: String, val yearPublished: Int)
+data class CollectionItemChangedEvent(
+        val collectionName: String,
+        val imageUrl: String,
+        val thumbnailUrl: String,
+        val yearPublished: Int
+)