+++ app/src/main/java/com/boardgamegeek/events/GameInfoChangedEvent.kt
@@ -1,10 +1,12 @@
 package com.boardgamegeek.events
 
-class GameInfoChangedEvent(
+data class GameInfoChangedEvent(
         val gameName: String,
         val subtype: String,
         val imageUrl: String,
         val thumbnailUrl: String,
+        val heroImageUrl: String,
         val arePlayersCustomSorted: Boolean,
-        val isFavorite: Boolean)
+        val isFavorite: Boolean
+)